+++ src/Pickles/Pickles.TestFrameworks.UnitTests/XUnit/XUnit1/WhenParsingxUnitResultsFile.cs
@@ -131,6 +131,12 @@ namespace PicklesDoc.Pickles.TestFrameworks.UnitTests.XUnit.XUnit1
         }
 
         [Test]
+        public new void ThenCanReadResultOfScenarioWithDanishCharacters()
+        {
+            base.ThenCanReadResultOfScenarioWithDanishCharacters();
+        }
+
+        [Test]
         public new void ThenCanReadResultOfScenarioOutlineWithAmpersand()
         {
           base.ThenCanReadResultOfScenarioOutlineWithAmpersand();