+++ test-harness/Cucumber/features/step_definitions/MinimalSteps.rb
@@ -1,4 +1,8 @@
-Then(/^passing step$/) do
+﻿Then(/^passing step$/) do
+  # Nothing to be done here
+end
+
+Then(/^the scenario with danish characters like æøå and ÆØÅ shall pass$/) do
   # Nothing to be done here
 end